exceptional handling for food and menu added

+++ src/main/java/com/fusemachine/controller/FoodController.java
@@ -1,6 +1,7 @@
 package com.fusemachine.controller;
 
 import com.fusemachine.entity.Food;
+import com.fusemachine.exceptions.ItemNotFoundException;
 import com.fusemachine.service.FoodService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -21,7 +22,11 @@ public class FoodController {
 
     @GetMapping("/foods/{id}")
     public Food findById(@PathVariable int id){
-        return foodService.findById(id);
+        Food food = foodService.findById(id);
+        if(food == null){
+            throw new ItemNotFoundException("Item with id = " + id + " not found.");
+        }
+        return food;
     }
 
     @PostMapping("/foods")
@@ -32,12 +37,18 @@ public class FoodController {
 
     @PutMapping("/foods/{id}")
     public void update(@RequestBody Food food, @PathVariable int id){
+        if(foodService.findById(id) == null){
+            throw new ItemNotFoundException("Item with id = " + id + " not found.");
+        }
         food.setId(id);
         foodService.save(food);
     }
 
     @DeleteMapping("/foods/{id}")
     public  void  deleteById(@PathVariable int id){
+        if(foodService.findById(id) == null){
+            throw new ItemNotFoundException("Item with id = " + id + " not found.");
+        }
         foodService.deleteById(id);
     }
 }
